4.21*: blogilistan laajennus, step9

+++ controllers/blogs.js
@@ -39,12 +39,22 @@ blogsRouter.delete('/:id', async (request, response) => {
         return response.status(400).json({ error: 'Invalid ID format' })
     }
 
-    const result = await Blog.findByIdAndDelete(id)
-    if (result) {
-        return response.status(204).end()
-    } else{
-      response.status(404).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+      return response.status(401).json({ error: 'token invalid' })
     }
+
+    const blog = await Blog.findById(id)
+    if(!blog){
+      return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if(blog.user.toString() !== decodedToken.id){
+      return response.status(401).json({ error: 'unauthorized user' })
+    }
+
+    await Blog.findByIdAndDelete(id)
+    return response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
